+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 These refer to the track2kml crate versions.
 
+## 0.3.0
+- Update COURAGEOUS format dependency to 0.4.0+schema.0.3.1
+
 ## 0.2.0
 - Update COURAGEOUS format dependency to 0.3.0